+++ muna/beta/metadata/tensorrt_rtx.py
@@ -7,7 +7,6 @@ from pydantic import Field
 from typing import Literal
 
 from ._torch import PyTorchInferenceMetadataBase
-from .tensorrt import TensorRTPrecision
 
 class TensorRTRTXInferenceMetadata(PyTorchInferenceMetadataBase):
     """
@@ -19,11 +18,6 @@ class TensorRTRTXInferenceMetadata(PyTorchInferenceMetadataBase):
         model_args (tuple): Positional inputs to the model.
         input_shapes (list): Model input tensor shapes. Use this to specify dynamic axes.
         optimum_config (optimum.ExporterConfig): Optimum exporter configuration. Required when `exporter` is `optimum`.
-        precision (TensorRTPrecision): TensorRT engine inference precision. Defaults to `fp16`.
+        kv_cache (KVCacheConfig): KV cache configuration for autoregressive models.
     """
     kind: Literal["meta.inference.tensorrt_rtx"] = Field(default="meta.inference.tensorrt_rtx", init=False)
-    precision: TensorRTPrecision = Field(
-        default="fp32",
-        description="TensorRT engine inference precision.",
-        exclude=True
-    )